+++ src/main/groovy/com/rapidminer/gradle/PluginConfiguration.groovy
@@ -23,7 +23,5 @@ package com.rapidminer.gradle;
  */
 public class PluginConfiguration {
 	String id
-	String description
-	String displayName
 	def tags = []
 }